Refactor optional fields validation into utils function
Also add notes to the optional fields validated in InstanceForm
